# build cached restmapper based on Kubernetes restmapper